NewickTreeTools

Documentation for NewickTreeTools.

Clustering.Hclust

NewickTreeTools.as_polytomy!

NewickTreeTools.collectiveLCA

NewickTreeTools.cuttree

NewickTreeTools.getleafnames

NewickTreeTools.levelorder

NewickTreeTools.majorityruletree

NewickTreeTools.mapinternalnodes

NewickTreeTools.maplocalnodes

NewickTreeTools.network_distance

NewickTreeTools.network_distances

NewickTreeTools.newick

NewickTreeTools.patristic_distance

NewickTreeTools.patristic_distances

NewickTreeTools.tally_tree_bifurcations

Clustering.Hclust — Method

Hclust(tree::Node)

convert ultrametric NewickTree.Node into Clustering.Hclust type

NewickTreeTools.as_polytomy! — Method

as_polytomy!(tree::Node; fun::Function=n->distance(n)≈0)

removes internal nodes from tree based on fun() which must return a Bool

by default removes zero length branches (i.e. nodes where distance between child and parent == 0)

NewickTreeTools.collectiveLCA — Method

collectiveLCA(tree, nodes)

finds last common ancester of a collection of Nodes

NewickTreeTools.cuttree — Method

cuttree(distfun::Function, tree::Node, θ)

returns all vector of Nodes where distance to root is greater than theta d > θ.

distance function must take to NewickTree.Node objects and compute a scaler distance between them.

NewickTreeTools.getleafnames — Method

getleafnames(t::Node)

Get names of all the leafs with t as ancester

NewickTreeTools.levelorder — Method

levelorder(v::CategoricalArray)

Get catagories from CategoricalArray as Vector{Int}

NewickTreeTools.majorityruletree — Method

Construct majority rule tree from tallylist (not well tested)

NewickTreeTools.mapinternalnodes — Method

mapinternalnodes(fun::Function, tree::Node, args...; kwargs...)

maps function fun() across internal nodes of tree.

args and kwargs are passed to fun()

NewickTreeTools.maplocalnodes — Method

maplocalnodes(fun::Function, tree::Node, args...; kwargs...)

maps function fun() across internal nodes of tree conditioned on having one direct child that is a leaf.

args and kwargs are passed to fun()

NewickTreeTools.network_distance — Method

network_distance(n::Node, m::Node)

shortest traversal distance between two nodes

sibling nodes would have network distance = 2

NewickTreeTools.network_distances — Method

network_distances(t::Node)

shortest traversal distance between all leafs with t as ancester sibling nodes would have network distance = 2

NewickTreeTools.newick — Method

newick(hc::Hclust; justroot=true, names=string.(hc.labels))

convert Hclust to newick tree; returns root node

Args:

hc, Hclust object from Clustering package

justroot, Bool return just the root or the vector of all nodes

names, Vector{String} names in same order as distance matrix

NewickTreeTools.patristic_distance — Method

patristic_distance(n::Node, m::Node)

shortest branch length path between two nodes.

sibling nodes i, and j of parent p would have patristic distance(i, p) + distance(j, p)

NewickTreeTools.patristic_distances — Method

patristic_distances(t::Node)

shortest branch length path between all leafs with t as ancester

sibling nodes i, and j of parent p would have patristic distance(i, p) + distance(j, p)

NewickTreeTools.tally_tree_bifurcations — Function

tally_tree_bifurcations(rootnode::AbstractTree, [cntr::Accumulator])

returns tally of leaf groups based on splitting tree at each internal node

assumes nodenames are strings parsible to integers, and returns Accumulator with key as a string: 1110000 where 1 = belonging to smaller group & 0 = belonging to the larger. values are the number of times this split is observed. cntr is modified in place so using the same counter in multiple calls will keep a tally across multiple trees